refactor(test-data): Consistent (more/less) filename
