+++ scripts/mock-exports.json
@@ -1,7 +1,7 @@
 [
-  "export { MockOntology } from \"./test/data/api/v2/mockOntology\";\n",
-  "export { MockList } from \"./test/data/api/v2/mockList\";\n",
+  "export { MockOntology } from \"./test/data/api/v2/mock-ontology\";\n",
+  "export { MockList } from \"./test/data/api/v2/mock-list\";\n",
   "export { MockResource } from \"./test/data/api/v2/mock-resource\";\n",
-  "export { MockProjects } from \"./test/data/api/admin/mockProjects\";\n",
-  "export { MockUsers } from \"./test/data/api/admin/mockUsers\";\n"
+  "export { MockProjects } from \"./test/data/api/admin/mock-projects\";\n",
+  "export { MockUsers } from \"./test/data/api/admin/mock-users\";\n"
 ]
