+++ src/cache/ListNodeV2Cache.spec.ts
@@ -1,7 +1,7 @@
 import { JsonConvert, OperationMode, ValueCheckingMode } from "json2typescript";
 import { PropertyMatchingRule } from "json2typescript/src/json2typescript/json-convert-enums";
 import { of } from "rxjs";
-import { MockList } from "../../test/data/api/v2/mockList";
+import { MockList } from "../../test/data/api/v2/mock-list";
 import { KnoraApiConfig } from "../knora-api-config";
 import { KnoraApiConnection } from "../knora-api-connection";
 import { ListNodeV2 } from "../models/v2/lists/list-node-v2";
