+++ src/models/v2/resources/ResourcesConversionUtil.spec.ts
@@ -2,8 +2,8 @@ import { JsonConvert, OperationMode, ValueCheckingMode } from "json2typescript";
 import { PropertyMatchingRule } from "json2typescript/src/json2typescript/json-convert-enums";
 import { of } from "rxjs";
 import { MockOntologyAssertions } from "../../../../test/data/api/v2/mock-ontology-assertions";
-import { MockList } from "../../../../test/data/api/v2/mockList";
-import { MockOntology } from "../../../../test/data/api/v2/mockOntology";
+import { MockList } from "../../../../test/data/api/v2/mock-list";
+import { MockOntology } from "../../../../test/data/api/v2/mock-ontology";
 import { KnoraApiConfig } from "../../../knora-api-config";
 import { KnoraApiConnection } from "../../../knora-api-connection";
 import { ReadResource } from "./read/read-resource";
